+++ scripts/macros/monsterFeatures/duergarMindMaster/duergarMindMaster.js
@@ -1,0 +1,6 @@
+import {mindMastery} from './mindMastery.js';
+import {reduce} from './reduce.js';
+export let duergarMindMaster = {
+    'reduce': reduce,
+    'mindMastery': mindMastery
+}